タイマー管理JSファイルの切り分け、タイマー表示ページの分岐実装
