+++ task_management/src/main/resources/templates/timers/breakTimer.html
@@ -21,8 +21,6 @@
 					<input type="hidden" name="settingId" th:value="${timersSetting.getId}" />
 					<p id="timerCount" th:text="${timersSetting.getBreakTime}"></p>	
 				</div>
-				<!-- タイマーの種類判定用に送信 -->
-				<input type="hidden" name="timerType" value="2">
 				<!-- スタートボタン -->
 				<input id="startButton"  type="submit" value="起動" />
 				<!-- ストップボタン -->
@@ -35,6 +33,6 @@
 	<!-- jqueryを読み込む -->
 	<script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
 	<!-- javascriptの読み込み -->	
-	<script src='/js/timerActions.js'></script>
+	<script src='/js/timerCountAction.js'></script>
 </body>
 </html>
