+++ task_management/src/main/resources/templates/timers/timerEdit.html
@@ -14,6 +14,9 @@
 			<input type="text" placeholder="00:00:00" th:field="*{focusTime}"/>
 			<label>休憩タイマー</label>
 			<input type="text" placeholder="00:00:00" th:field="*{breakTime}"/>
+			<input type="hidden" th:field="*{registarAt}">
+			<!-- ページ遷移先判定用のデータ送信 -->
+			<input type="hidden" name="formSendCheckPageValue" value="2">
 			<div id="select-button">
 				<button type="submit">登録</button>
 				<a th:href="@{/getTimerlist}">キャンセル</a>
