+++ task_management/src/main/resources/templates/timers/timers.html
@@ -42,5 +42,5 @@
 <!-- jquetyを読み込む -->
 <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
 <!-- JavaScriptを読み込む -->
-<script src='/js/timerActions.js'></script>
+<script src='/js/timerEditAction.js'></script>
 </html>
